Working Dataset prototype for TrackML
Implements the following classes:
- `TrackMLDataset`
- `WedgePatchify3d`

`TrackMLDataset` is designed to be almost a drop-in replacement for
`toytrack.TracksDataset`, at least in terms of API and for the default
parameters. The scale and multiplicity of data is expected to be
significantly different. It is not optimised for speed, since it
operates on the full, unskimmed HDF5 file, and performs the inner join
on hit_id on the fly.

Additionally, add extra validation checks to a previous notebook.